Moving classes with single method to other places and deleting unused files
