+++ bayesian_cnn_prometheus/preprocessing/data_generator.py
@@ -1,11 +1,11 @@
 import functools
-from typing import Dict
+from typing import Dict, Generator
 
 import numpy as np
+from bayesian_cnn_prometheus.preprocessing.preprocessing_links import ImageLoader
 
 from bayesian_cnn_prometheus.constants import DatasetType
 from bayesian_cnn_prometheus.preprocessing.data_splitter import DataSplitter
-from bayesian_cnn_prometheus.preprocessing.preprocessing_links import ImageLoader, ImageNormalizer, ChunksGenerator
 
 
 class DataGenerator:
@@ -46,10 +46,10 @@ class DataGenerator:
         """
         for image_index in self.dataset_structure[dataset_type]:
             x_npy, y_npy = self.image_loader.load(image_index)
-            x_npy_norm = ImageNormalizer.normalize(x_npy) if self.should_normalise else x_npy
+            x_npy_norm = DataGenerator.normalize(x_npy) if self.should_normalise else x_npy
             images_chunks, targets_chunks = [], []
-            for x_chunk, y_chunk in zip(ChunksGenerator.generate(x_npy_norm, self.chunk_size),
-                                        ChunksGenerator.generate(y_npy, self.chunk_size)):
+            for x_chunk, y_chunk in zip(DataGenerator._generate_chunks(x_npy_norm, self.chunk_size),
+                                        DataGenerator._generate_chunks(y_npy, self.chunk_size)):
                 x_chunk = x_chunk.reshape((*x_chunk.shape, 1))
                 y_chunk = y_chunk.reshape((*y_chunk.shape, 1))
 
@@ -58,3 +58,29 @@ class DataGenerator:
 
                 if len(images_chunks) == batch_size and len(targets_chunks) == batch_size:
                     yield np.array(images_chunks), np.array(targets_chunks)
+
+    @staticmethod
+    def _generate_chunks(data_subset: np.array, chunk_size: tuple = (32, 32, 16)) -> Generator[np.ndarray, None, None]:
+        """
+        Generates chunks from the original data (numpy array).
+        :param data_subset: single subset of data (or labels)
+        :param chunk_size: size of 3d chunk (a, b, c) to train the model with them
+        :return: generator which produces chunks with size (a, b, c)
+        """
+        origin_x, origin_y, origin_z = data_subset.shape
+        chunk_x, chunk_y, chunk_z = chunk_size
+
+        for x in range(0, origin_x, chunk_x)[:-1]:
+            for y in range(0, origin_y, chunk_y)[:-1]:
+                for z in range(0, origin_z, chunk_z)[:-1]:
+                    chunk = data_subset[x:x + chunk_x, y:y + chunk_y, z:z + chunk_z]
+                    yield chunk
+
+    @staticmethod
+    def normalize(image: np.array) -> np.array:
+        """
+        Transforms data to have mean 0 and std 1 (standardize).
+        :param image: non-standardized image to transform
+        :return standardized image
+        """
+        return (image - np.mean(image)) / np.std(image)
