+++ bayesian_cnn_prometheus/preprocessing/image_loader.py
