--- bayesian_cnn_prometheus/preprocessing/preprocessing_links/image_normalizer.py
@@ -1,13 +1,0 @@
-import numpy as np
-
-
-class ImageNormalizer:
-
-    @staticmethod
-    def normalize(image: np.array) -> np.array:
-        """
-        Transforms data to have mean 0 and std 1 (standardize).
-        :param image: non-standardized image to transform
-        :return standardized image
-        """
-        return (image - np.mean(image)) / np.std(image)
